Split into Rust and GEOS namespaces
