+++ py-geopolars/python/geopolars/internals/georust/__init__.py
@@ -1,0 +1,1 @@
+from .geoseries import GeoRustSeriesOperations
