+++ py-geopolars/python/geopolars/internals/geos/__init__.py
@@ -1,0 +1,1 @@
+from .geoseries import GEOSSeriesOperations
